still working on video, added real arrows to page tho (forgot style lol)
