+++ package.json
@@ -33,7 +33,7 @@
     "react-native-safe-area-context": "4.12.0",
     "react-native-screens": "~4.0.0",
     "react-native-slider": "^0.11.0",
-    "react-native-video": "^6.6.4",
+    "react-native-video": "^6.8.0",
     "react-native-video-controls": "^2.8.1"
   },
   "devDependencies": {
